+++ dbqm/ui/widgets/__init__.py
@@ -1,6 +1,5 @@
 """Reusable UI widgets for the DBQM Textual application."""
 
-from dbqm.ui.widgets.breadcrumb import Breadcrumb, BreadcrumbNavigated
 from dbqm.ui.widgets.status_bar import StatusBar
 from dbqm.ui.widgets.action_bar import ActionBar, ActionSelected, Action
 from dbqm.ui.widgets.result_table import ResultTable
@@ -12,8 +11,6 @@ from dbqm.ui.widgets.dialog import Dialog
 from dbqm.ui.widgets.empty_state import EmptyState
 
 __all__ = [
-    "Breadcrumb",
-    "BreadcrumbNavigated",
     "StatusBar",
     "ActionBar",
     "ActionSelected",